requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
python-dotenv==1.0.0
byteplussdkarkruntime==1.0.12
google-search-results==2.4.2
Pillow==10.1.0
opencv-python==4.8.1.78
requests==2.31.0